+++ tracklytics-runtime/src/main/java/com/orhanobut/tracklytics/trackers/SimpleTrackingAdapter.java
@@ -1,13 +1,13 @@
 package com.orhanobut.tracklytics.trackers;
 
-import com.orhanobut.tracklytics.TrackEvent;
+import com.orhanobut.tracklytics.Event;
 
 import java.util.Map;
 
 public class SimpleTrackingAdapter implements TrackingAdapter {
 
-  @Override public void trackEvent(TrackEvent event, Map<String, Object> attributes,
-                                   Map<String, Object> superAttributes) {
+  @Override public void trackEvent(Event event, Map<String, Object> superAttributes) {
+
   }
 
   @Override public void start() {